refactor: delete exception folder in the exception package
