+++ src/main/java/com/kopchak/worldoftoys/exception/CategoryException.java
@@ -1,4 +1,4 @@
-package com.kopchak.worldoftoys.exception.exception;
+package com.kopchak.worldoftoys.exception;
 
 public class CategoryException extends Exception {
     public CategoryException(String reason) {
