+++ src/main/java/com/kopchak/worldoftoys/exception/ImageDecompressionException.java
@@ -1,4 +1,4 @@
-package com.kopchak.worldoftoys.exception.exception;
+package com.kopchak.worldoftoys.exception;
 
 public class ImageDecompressionException extends Exception{
     public ImageDecompressionException(String message) {
