+++ src/main/java/com/kopchak/worldoftoys/exception/ImageExceedsMaxSizeException.java
@@ -1,4 +1,4 @@
-package com.kopchak.worldoftoys.exception.exception;
+package com.kopchak.worldoftoys.exception;
 
 public class ImageExceedsMaxSizeException extends Exception{
     public ImageExceedsMaxSizeException(String message) {
