+++ src/main/java/com/kopchak/worldoftoys/exception/InvalidConfirmationTokenException.java
@@ -1,4 +1,4 @@
-package com.kopchak.worldoftoys.exception.exception;
+package com.kopchak.worldoftoys.exception;
 
 public class InvalidConfirmationTokenException extends Exception{
     public InvalidConfirmationTokenException(String reason) {
