+++ src/main/java/com/kopchak/worldoftoys/exception/InvalidOrderException.java
@@ -1,4 +1,4 @@
-package com.kopchak.worldoftoys.exception.exception;
+package com.kopchak.worldoftoys.exception;
 
 public class InvalidOrderException extends Exception{
     public InvalidOrderException(String message) {
