+++ src/main/java/com/kopchak/worldoftoys/exception/MessageSendingException.java
@@ -1,4 +1,4 @@
-package com.kopchak.worldoftoys.exception.exception;
+package com.kopchak.worldoftoys.exception;
 
 public class MessageSendingException extends Exception {
     public MessageSendingException(String message) {
