+++ src/main/java/com/kopchak/worldoftoys/exception/ProductNotFoundException.java
@@ -1,4 +1,4 @@
-package com.kopchak.worldoftoys.exception.exception;
+package com.kopchak.worldoftoys.exception;
 
 public class ProductNotFoundException extends Exception{
     public ProductNotFoundException(String reason) {
