+++ src/main/java/com/kopchak/worldoftoys/exception/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.kopchak.worldoftoys.exception.exception;
+package com.kopchak.worldoftoys.exception;
 
 public class UserNotFoundException extends Exception{
     public UserNotFoundException(String reason) {
