+++ src/main/java/com/kopchak/worldoftoys/exception/UsernameAlreadyExistException.java
@@ -1,4 +1,4 @@
-package com.kopchak.worldoftoys.exception.exception;
+package com.kopchak.worldoftoys.exception;
 
 public class UsernameAlreadyExistException extends Exception{
     public UsernameAlreadyExistException(String reason) {
