+++ src/main/java/com/kopchak/worldoftoys/exception/handler/GlobalExceptionHandler.java
@@ -2,7 +2,7 @@ package com.kopchak.worldoftoys.exception.handler;
 
 import com.kopchak.worldoftoys.dto.error.MethodArgumentNotValidExceptionDto;
 import com.kopchak.worldoftoys.dto.error.ResponseStatusExceptionDto;
-import com.kopchak.worldoftoys.exception.exception.JwtTokenException;
+import com.kopchak.worldoftoys.exception.JwtTokenException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.FieldError;
