+++ src/main/java/com/kopchak/worldoftoys/filter/JwtAuthenticationFilter.java
@@ -1,6 +1,6 @@
 package com.kopchak.worldoftoys.filter;
 
-import com.kopchak.worldoftoys.exception.exception.JwtTokenException;
+import com.kopchak.worldoftoys.exception.JwtTokenException;
 import com.kopchak.worldoftoys.model.token.AuthTokenType;
 import com.kopchak.worldoftoys.service.JwtTokenService;
 import jakarta.servlet.FilterChain;
