+++ src/main/java/com/kopchak/worldoftoys/mapper/order/OrderMapper.java
@@ -4,7 +4,7 @@ import com.kopchak.worldoftoys.dto.admin.product.order.FilteredOrdersPageDto;
 import com.kopchak.worldoftoys.dto.admin.product.order.FilteringOrderOptionsDto;
 import com.kopchak.worldoftoys.dto.admin.product.order.StatusDto;
 import com.kopchak.worldoftoys.dto.order.OrderDto;
-import com.kopchak.worldoftoys.exception.exception.OrderCreationException;
+import com.kopchak.worldoftoys.exception.OrderCreationException;
 import com.kopchak.worldoftoys.model.cart.CartItem;
 import com.kopchak.worldoftoys.model.order.Order;
 import com.kopchak.worldoftoys.model.order.OrderStatus;
