+++ src/main/java/com/kopchak/worldoftoys/repository/product/ProductCategoryRepository.java
@@ -1,7 +1,7 @@
 package com.kopchak.worldoftoys.repository.product;
 
 import com.kopchak.worldoftoys.dto.product.category.FilteringProductCategoriesDto;
-import com.kopchak.worldoftoys.exception.exception.CategoryException;
+import com.kopchak.worldoftoys.exception.CategoryException;
 import com.kopchak.worldoftoys.model.product.Product;
 import com.kopchak.worldoftoys.model.product.category.ProductCategory;
 import org.springframework.data.jpa.domain.Specification;
