+++ src/main/java/com/kopchak/worldoftoys/service/JwtTokenService.java
@@ -2,8 +2,8 @@ package com.kopchak.worldoftoys.service;
 
 import com.kopchak.worldoftoys.dto.token.AccessAndRefreshTokensDto;
 import com.kopchak.worldoftoys.dto.token.AuthTokenDto;
-import com.kopchak.worldoftoys.exception.exception.JwtTokenException;
-import com.kopchak.worldoftoys.exception.exception.TokenAlreadyExistException;
+import com.kopchak.worldoftoys.exception.JwtTokenException;
+import com.kopchak.worldoftoys.exception.TokenAlreadyExistException;
 import com.kopchak.worldoftoys.model.token.AuthTokenType;
 import com.kopchak.worldoftoys.model.user.AppUser;
