+++ src/main/java/com/kopchak/worldoftoys/service/PaymentService.java
@@ -1,8 +1,8 @@
 package com.kopchak.worldoftoys.service;
 
 import com.kopchak.worldoftoys.dto.payment.StripeCredentialsDto;
-import com.kopchak.worldoftoys.exception.exception.InvalidOrderException;
-import com.kopchak.worldoftoys.exception.exception.MessageSendingException;
+import com.kopchak.worldoftoys.exception.InvalidOrderException;
+import com.kopchak.worldoftoys.exception.MessageSendingException;
 import com.stripe.exception.StripeException;
 
 public interface PaymentService {
