+++ src/main/java/com/kopchak/worldoftoys/service/UserService.java
@@ -3,10 +3,10 @@ package com.kopchak.worldoftoys.service;
 import com.kopchak.worldoftoys.dto.token.AccessAndRefreshTokensDto;
 import com.kopchak.worldoftoys.dto.user.UserAuthDto;
 import com.kopchak.worldoftoys.dto.user.UserRegistrationDto;
-import com.kopchak.worldoftoys.exception.exception.AccountActivationException;
-import com.kopchak.worldoftoys.exception.exception.InvalidPasswordException;
-import com.kopchak.worldoftoys.exception.exception.UserNotFoundException;
-import com.kopchak.worldoftoys.exception.exception.UsernameAlreadyExistException;
+import com.kopchak.worldoftoys.exception.AccountActivationException;
+import com.kopchak.worldoftoys.exception.InvalidPasswordException;
+import com.kopchak.worldoftoys.exception.UserNotFoundException;
+import com.kopchak.worldoftoys.exception.UsernameAlreadyExistException;
 import com.kopchak.worldoftoys.model.user.AppUser;
 
 public interface UserService {
