+++ src/main/java/com/kopchak/worldoftoys/service/impl/ImageServiceImpl.java
@@ -1,10 +1,10 @@
 package com.kopchak.worldoftoys.service.impl;
 
 import com.kopchak.worldoftoys.dto.image.ImageDto;
-import com.kopchak.worldoftoys.exception.exception.ImageCompressionException;
-import com.kopchak.worldoftoys.exception.exception.ImageDecompressionException;
-import com.kopchak.worldoftoys.exception.exception.ImageExceedsMaxSizeException;
-import com.kopchak.worldoftoys.exception.exception.InvalidImageFileFormatException;
+import com.kopchak.worldoftoys.exception.ImageCompressionException;
+import com.kopchak.worldoftoys.exception.ImageDecompressionException;
+import com.kopchak.worldoftoys.exception.ImageExceedsMaxSizeException;
+import com.kopchak.worldoftoys.exception.InvalidImageFileFormatException;
 import com.kopchak.worldoftoys.model.image.Image;
 import com.kopchak.worldoftoys.model.product.Product;
 import com.kopchak.worldoftoys.repository.product.image.ImageRepository;
